requirements.txt
# Core
openai==1.109.1
python-dotenv==1.0.1

# LangChain family (all updated to latest compatible versions)
langchain==0.3.13
langchain-core==0.3.79
langchain-community==0.3.13
langchain-openai==0.2.9
langchain-text-splitters==0.3.2

# PDF + Vector store
pypdf==4.3.1
faiss-cpu==1.12.0

# UI
gradio==4.44.1
gradio-client==1.3.0
